Move tokenization code into its own module, down in 'lutris.util'.
